refactor: function to sync & subscribe addresses
