+++ jobs/workers.ts
@@ -1,38 +1,11 @@
 import { Worker, Job, Queue } from 'bullmq'
-import { Address, Prisma } from '@prisma/client'
 import { redisBullMQ } from 'redis/clientInstance'
 import { SYNC_NEW_ADDRESSES_DELAY, DEFAULT_WORKER_LOCK_DURATION, RESPONSE_MESSAGES, KeyValueT } from 'constants/index'
 
 import * as transactionService from 'services/transactionService'
 import * as priceService from 'services/priceService'
 import * as addressService from 'services/addressService'
-import { subscribeAddressesAddTransactions } from 'services/blockchainService'
 import { parseError } from 'utils/validators'
-import { productionAddresses } from 'prisma/seeds/addresses'
-import { appendTxsToFile } from 'prisma/seeds/transactions'
-
-const syncAndSubscribeAddresses = async (addresses: Address[]): Promise<KeyValueT<string>> => {
-  const failedAddressesWithErrors: KeyValueT<string> = {}
-  let txsToSave: Prisma.TransactionCreateManyInput[] = []
-  const productionAddressesIds = productionAddresses.map(addr => addr.id)
-  await Promise.all(
-    addresses.map(async (addr) => {
-      try {
-        await subscribeAddressesAddTransactions([addr])
-        const txs = await transactionService.syncAllTransactionsForAddress(addr, Infinity)
-        if (productionAddressesIds.includes(addr.id)) {
-          txsToSave = txsToSave.concat(txs)
-        }
-      } catch (err: any) {
-        failedAddressesWithErrors[addr.address] = err.stack
-      }
-    })
-  )
-  if (txsToSave.length !== 0) {
-    await appendTxsToFile(txsToSave)
-  }
-  return failedAddressesWithErrors
-}
 
 const syncAndSubscribeAllAddressTransactionsForNetworkJob = async (job: Job): Promise<void> => {
   console.log(`job ${job.id as string}: syncing and subscribing all addresses for network ${job.data.networkId as string}...`)
@@ -40,7 +13,7 @@ const syncAndSubscribeAllAddressTransactionsForNetworkJob = async (job: Job): Pr
   try {
     let addresses = await addressService.fetchAllAddressesForNetworkId(job.data.networkId)
     addresses = addresses.filter(addr => addr.lastSynced == null)
-    failedAddressesWithErrors = await syncAndSubscribeAddresses(addresses)
+    failedAddressesWithErrors = await transactionService.syncAndSubscribeAddresses(addresses)
   } catch (err: any) {
     const parsedError = parseError(err)
     if (parsedError.message === RESPONSE_MESSAGES.TRANSACTION_ALREADY_EXISTS_FOR_ADDRESS_400.message) {
